+++ olympia_engine/src/remote/loader.rs
@@ -1,0 +1,9 @@
+pub enum DataType {
+    ROM,
+}
+
+pub trait DataManager {
+    type Error;
+    type Identifier;
+    fn load(&self, ty: DataType, identifier: Self::Identifier) -> Result<Vec<u8>, Self::Error>;
+}