+++ olympia_core/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.3.0
+
+* Add some convenience methods for use in frontends
+* Sync version with olympia_engine
+
 # 0.1.0
 
 * Initial implementation