+++ olympia_derive/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.3.0
+
+* Replace `thiserror` with `derive_more`
+* Sync version with `olympia_engine`
+
 ## 0.1.0
 
 * Initial version