+++ olympia_engine/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 0.3.0
+
+### Added features
+
+* Add support for PPU background tiles
+* Add support for remote usage
+* Provide `std::error::Error` implementations for all errors when  the `std` feature is enabled
+* Add new event handling system to monitor local events
+
+### Breaking Changes
+
+* Instructions have been totally rewritten. The old `Instruction` type has been replaced with `RuntimeInstruction`
+* Many types that are needed for new derives have been moved to the olympia_core crate. In most cases these should
+  be re-exported with their old names.
+* Renames:
+  * `rom::CartridgeEnum` -> `rom::ControllerEnum`
+  * `rom::CartridgeType` -> `rom::CartridgeController`
+
 ## 0.2.0
 
 ### Added features